Move DraggableItemTouchHelper to widget package
